requirements.txt
kafka-python-ng
pandas
gpsd-py3